requirements.txt
mysql-connector-python == 8.0.23
pandas == 0.24.2   
SQLAlchemy == 1.2.18
